added travis config and unit test coverage (switched of, waitn for more tests)
