+++ .gitignore
@@ -4,3 +4,5 @@ test/e2e/generated
 src/test/generated
 lib
 node_modules
+./coverage
+./.nyc_output
