+++ package.json
@@ -12,14 +12,18 @@
   "scripts": {
     "build": "rimraf lib && tsc && node scripts/copy-templates.js",
     "watch": "tsc -w",
-    "test": "mocha --config test/mocha.js.opts",
-
+    "test": "mocha",
+    "coverage": "nyc npm run test",
     "test:e2e": "mocha 'test/e2e/*.test.js' -g 'test:e2e:polymer2|test:e2e:sdk|test:e2e:react' --timeout 0",
     "test:e2e:sdk": "mocha 'test/e2e/sdk.test.js'  -g 'test:e2e:sdk' --timeout 0",
-    "test:e2e:react": "mocha 'test/e2e/react-client.test.js'  -g 'test:e2e:react' --timeout 0",
+
+    "test:e2e:react": "mocha 'test/e2e/react-client--*.test.js'  -g 'test:e2e:react' --timeout 0",
+    "test:e2e:react:mpg": "mocha 'test/e2e/react-client--mpg.test.js'  -g 'test:e2e:react:mpg' --timeout 0",
+    "test:e2e:react:empty-model": "mocha 'test/e2e/react-client--empty-model.test.js'  -g 'test:e2e:react:empty-model' --timeout 0",
+    "test:e2e:react:trusty-dragon": "mocha 'test/e2e/react-client--trusty-dragon.test.js'  -g 'test:e2e:react:trusty-dragon' --timeout 0",
+
     "test:e2e:polymer2": "mocha 'test/e2e/polymer2.test.js'  -g 'test:e2e:polymer2' --timeout 0",
     "test:e2e:polymer2-ts": "mocha 'test/e2e/polymer2-ts.test.js'  -g 'test:e2e:polymer2-ts' --timeout 0",
-
     "prepublishOnly": "npm run build"
   },
   "repository": {
@@ -34,16 +38,33 @@
     "yeoman-generator": "~4.0.1"
   },
   "devDependencies": {
+    "@istanbuljs/nyc-config-typescript": "^0.1.3",
     "@types/ejs": "^2.5.1",
     "@types/mocha": "^2.2.48",
     "@types/node": "^9.6.17",
     "@types/through2": "^2.0.33",
     "@types/yeoman-generator": "^2.0.3",
     "mocha": "^5.1.1",
+    "nyc": "^14.1.1",
     "rimraf": "^2.6.2",
     "source-map-support": "^0.5.6",
     "ts-node": "^4.1.0",
     "vinyl-fs": "^3.0.3"
+  },
+  "nyc": {
+    "extends": "@istanbuljs/nyc-config-typescript",
+    "all": true,
+    "include": [
+      "src/**/*.ts"
+    ],
+    "exclude": [
+      "src/test/**",
+      "src/generators/polymer2/**",
+      "src/generators/polymer2-typescript/**",
+      "src/generators/app/template/**"
+    ],
+    "reporter":["text", "html"],
+    "check-coverage": true
   },
   "author": "CUBA Platform Team <info@cuba-platform.com>",
   "license": "Apache-2.0"
