--- test/mocha.js.opts
@@ -1,2 +1,0 @@
---require source-map-support/register
-lib/test/**/*.test.js
