+++ test/mocha.opts
@@ -1,3 +1,2 @@
 --require ts-node/register
 src/test/**/*.test.ts
---timeout 0
